Use socket specs for getaddrinfo() in sock_connect()
In BaseSelectorEventLoop.sock_connect(), an address is resolved if needed, but
getaddrinfo() isn't called with the socket's parameters (family, proto, etc).
If the resolved address mismatch the socket type, the connection fails.

For instance, if getaddrinfo() returns an IPv6 (AF_INET6) resolution and the
socket is an AF_INET, sock_connect() eventually raises a TypeError.