Don't import tty module on Windows
While the Windows release of Python seems to include the 'tty' module,
attempting to import it fails with an error about not finding the
'termios' module. Since the tty tests here aren't run on Windows,
though, there's no need to import tty there. This commit makes the
import conditional on the platform.