Don't select for read on character devices in _UnixWritePipeTransport (#374)
* Don't select for read on character devices in _UnixWritePipeTransport

_UnixWritePipeTransport selects for read on write pipes to detect when
the remote end of the pipe is closed. This works for unidirectional FIFOs
and dedicated socket pairs where nothing is written to the read side of
the pair, but it can cause problems with devices or sockets where
bidirectional I/O is being done.

This commit changes _UnixWritePipeTransport to only select for read on
sockets and FIFOs (on OSes which support that), and not on character
devices. When connect_write_pipe() is used on those devices,
end-of-file will need to be detected using a read pipe which accepts
inputs from the device.

No change is made to how sockets are handled, so passing in a socket
used for bidirectional I/O to connect_write_pipe() is not supported.
Async I/O on sockets should be performed using functions like
BaseEventLoop.create_connection(). Socket pairs which are only
used for undirectional I/O can be used here, though.

* Add new unit test for bidirectional I/O on TTYs

* Don't import tty module on Windows

While the Windows release of Python seems to include the 'tty' module,
attempting to import it fails with an error about not finding the
'termios' module. Since the tty tests here aren't run on Windows,
though, there's no need to import tty there. This commit makes the
import conditional on the platform.