Accept multiple connections on listener EVENT_READ
When a large number of connections are made at once the event loop
struggles to accept them because it only accepts and begins
handling one new connection on each pass of the loop in which the
EVENT_READ is emitted for the listening socket. This causes
incoming connections to time out and sever.

This patch uses the backlog setting of the listening socket to
allow for accepting a bounded number of new incoming connections
rather than a single one.

Patch by Kevin Conway <kevinjacobconway@gmail.com>.